kafka-producers/requirements.txt
kafka-python>=2.0.2
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
